add share, latency and http resolver options, refactor error response
